% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_bottom_weight_layer}
\alias{compute_bottom_weight_layer}
\title{Compute bottom-only layer weights}
\usage{
compute_bottom_weight_layer(bgm_file, nc_mask_path)
}
\arguments{
\item{bgm_file}{Path to the BGM file.}

\item{nc_mask_path}{Path to a reference NetCDF file for the NA mask.}
}
\value{
A numeric array of dimension \link{7, 89, 730}.
}
\description{
Sets layer 1 (bottom) to box area and all other layers to NA, masking open
ocean boxes (61:89). Used by the "bottom" layer mode.
}
\keyword{internal}
