% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radar_data.R
\name{get_radar_data}
\alias{get_radar_data}
\title{Code to create a radar plot for Vital Sign Indicators}
\usage{
get_radar_data(
  this.scenario,
  file.name,
  startyear,
  runtime,
  outputfrequency,
  functionalgroup.file,
  this.output.nc,
  atlantis.table,
  folder.base,
  biomass.paths
)
}
\description{
https://vitalsigns.pugetsoundinfo.wa.gov/
}
\author{
Hem Nalini Morzaria Luna
}
